wip:리스트 컴포넌트에서 소득 타입,금액 입력시 현재의 금액 상황을 보여주는 컴포넌트 생성. 함수 구현 완료
